lesson 7

+++ model/contacts.js
@@ -1,13 +1,19 @@
 const Contacts = require('./Schemas/contact')
 
 
-const listContacts = async () => {
-  const results = await Contacts.find({})
+const listContacts = async (userId) => {
+  const results = await Contacts.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'name email sex - id'
+  })
   return results
 }
 
-const getContactById = async (id) => {
-  const result = await Contacts.findOne({ _id: id })
+const getContactById = async (id, userId) => {
+  const result = await Contacts.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: 'name email sex - id'
+  })
   return result
 }
 const addContact = async (body) => {
@@ -15,14 +21,14 @@ const addContact = async (body) => {
   return result
 }
 
-const removeContact = async (id) => {
-  const result = await Contacts.findByIdAndRemove({ _id: id })
+const removeContact = async (id, userId) => {
+  const result = await Contacts.findByIdAndRemove({ _id: id, owner: userId })
   return result
 }
 
-const updateContact = async (id, body) => {
+const updateContact = async (id, body, userId) => {
   const result = await Contacts.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body },
     { new: true }
   )
